requirements.txt
Django
django-environ
django-role-permissions
djangorestframework
djangorestframework-simplejwt
mysqlclient